+++ Changelog.md
@@ -3,3 +3,5 @@
  - Moved plotting code into a libary file to unclutter the notebooks
  - Cleaned up Scripts for deployment and execution
  - Documented all steps nassesary for a repoduction in [Nodes.md](Nodes.md)
+ - Added new Function 2.0 (async) implementation for Azure Functions
+ - Relaxed timeout constraints from 30 to 60 seconds for MAF Workloads